Refactor/general

+++ src/sim/core/events.rs
@@ -1,5 +1,4 @@
-use crate::neural::AgentId;
-use crate::sim::storage::CommunityId;
+use crate::core::{AgentId, CommunityId};
 use crate::vm::{ByteStack, VmContext, op};
 
 pub struct SimulationContext {
@@ -54,7 +53,7 @@ impl VmContext for SimulationContext {
                 let target_id = stack.pop();
                 self.events.push(SimulationEvent::LeaveCommunity {
                     agent_id: self.agent_id,
-                    target_community_id: CommunityId(target_id as usize),
+                    target_community_id: CommunityId(target_id as u64),
                 });
                 true
             }
@@ -76,7 +75,7 @@ impl VmContext for SimulationContext {
 mod tests {
     use super::*;
     use crate::vm::ByteStack;
-    use crate::vm::isa::op;
+    use crate::vm::op;
 
     #[test]
     fn test_leave_community_syscall() {
